following norm changes
