+++ Makefile
@@ -6,7 +6,7 @@
 #    By: mely-pan <mely-pan@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/05/31 20:14:42 by mely-pan          #+#    #+#              #
-#    Updated: 2025/05/31 20:14:43 by mely-pan         ###   ########.fr        #
+#    Updated: 2025/05/31 22:07:38 by mely-pan         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -22,7 +22,8 @@ SRC = ./src/free.c ./src/get_env.c ./src/init_ms.c ./src/handle/env_var.c \
 	./src/execute/executor.c ./src/execute/executor_utils.c \
 	./src/execute/executor_ops.c ./src/builtins/dir.c ./src/builtins/echo.c \
 	./src/builtins/export.c ./src/builtins/unset.c ./src/execute/redir.c \
-	./src/execute/redir_utils.c ./src/main.c ./src/free_utils.c ./src/debug.c 
+	./src/execute/redir_utils.c ./src/main.c ./src/free_utils.c ./src/debug.c \
+	./src/execute/redir_utils2.c
 
 OBJ_DIR = ./obj
 OBJ = $(patsubst ./src/%.c, $(OBJ_DIR)/%.o, $(SRC))
