+++ src/execute/redir.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   redirection.c                                      :+:      :+:    :+:   */
+/*   redir.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: htrindad <htrindad@student.42lisboa.com>   +#+  +:+       +#+        */
+/*   By: mely-pan <mely-pan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/19 20:34:14 by mely-pan          #+#    #+#             */
-/*   Updated: 2025/05/26 20:11:34 by htrindad         ###   ########.fr       */
+/*   Updated: 2025/05/31 22:06:16 by mely-pan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	add_redir(t_redir **redir_list, t_case type, char *filename)
 {
-	t_redir *tmp;
+	t_redir	*tmp;
 	t_redir	*new;
 
 	new = ft_calloc(1, sizeof(t_redir));
@@ -37,42 +37,6 @@ int	add_redir(t_redir **redir_list, t_case type, char *filename)
 	return (0);
 }
 
-int	handle_redirections(t_token *tokens)
-{
-	t_redir	*redir;
-	int		fd;
-
-	redir = tokens->fds->in;
-	while (redir)
-	{
-		if (redir->type == IN)
-			fd = open(redir->filename, O_RDONLY);
-		else if (redir->type == HEREDOC)
-			fd = handle_heredoc(redir->filename);
-		if (fd < 0)
-			return (perror(redir->filename), 1);
-		if (dup2(fd, STDIN_FILENO) < 0)
-			return (perror("dup2"), close(fd), 1);
-		close(fd);
-		redir = redir->next;
-	}
-	redir = tokens->fds->out;
-	while (redir)
-	{
-		if (redir->type == OUT)
-			fd = open(redir->filename, O_WRONLY | O_CREAT | O_TRUNC, 0644);
-		else
-			fd = open(redir->filename, O_WRONLY | O_CREAT | O_APPEND, 0644);
-		if (fd < 0)
-			return (perror(redir->filename), 1);
-		if (dup2(fd, STDOUT_FILENO) < 0)
-			return (perror("dup2"), close(fd), 1);
-		close(fd);
-		redir = redir->next;
-	}
-	return (0);
-}
-
 bool	is_redirection(t_case type)
 {
 	if (type != NONE && type != PIPE)
@@ -89,12 +53,14 @@ static int	set_redir(t_token *tok)
 	{
 		if (tok->cchar == IN || tok->cchar == HEREDOC)
 		{
-			if (tmp->next && tmp->next->value && add_redir(&tok->fds->in, tmp->cchar, tmp->next->value[0]))
+			if (tmp->next && tmp->next->value && add_redir(&tok->fds->in,
+					tmp->cchar, tmp->next->value[0]))
 				return (1);
 		}
 		else
 		{
-			if (tmp->next && tmp->next->value && add_redir(&tok->fds->out, tmp->cchar, tmp->next->value[0]))
+			if (tmp->next && tmp->next->value && add_redir(&tok->fds->out,
+					tmp->cchar, tmp->next->value[0]))
 				return (1);
 		}
 		tmp = tmp->next;
@@ -102,43 +68,37 @@ static int	set_redir(t_token *tok)
 	return (0);
 }
 
+static void	mark_chained_redirs(t_token **curr)
+{
+	while ((*curr)->next && is_redirection((*curr)->next->cchar))
+	{
+		(*curr)->next->is_redir = true;
+		(*curr)->next->fds = NULL;
+		(*curr) = (*curr)->next;
+	}
+}
+
 int	parse_redirections(t_token **tokens)
 {
 	t_token	*curr;
 
 	curr = *tokens;
-	if (DEBUG)
-		print_tokens(*tokens);
 	while (curr)
 	{
-		if (curr && is_redirection(curr->cchar) && curr->is_redir == false)
+		if (is_redirection(curr->cchar) && !curr->is_redir)
 		{
 			if (!curr->next || !curr->next->value || !curr->next->value[0])
 				return (1);
-			if (!curr->fds)
-			{
-				curr->fds = ft_calloc(1, sizeof(t_fds));
-				if (!curr->fds)
-					return (1);
-				curr->fds->in = NULL;
-				curr->fds->out = NULL;
-			}
+			if (alloc_fds_if_needed(curr))
+				return (1);
 			if (set_redir(curr))
 				return (1);
-			while (curr->next && is_redirection(curr->next->cchar))
-			{
-				curr->next->is_redir = true;
-				curr->next->fds = NULL;
-				curr = curr->next;
-			}
+			mark_chained_redirs(&curr);
 		}
 		else
 			curr->fds = NULL;
-		if (curr)
-			curr = curr->next;
+		curr = curr->next;
 	}
 	cleanup_redir(tokens);
-	if (DEBUG)
-		print_tokens_debug(*tokens);
 	return (0);
 }
